requirements.txt
pandas==2.2.3
requests==2.32.3
pydantic==2.10.6
pyyaml==6.0.2
sqlalchemy==2.0.36
